+++ dotfiles/.local/share/gnome-shell/extensions/just-perfection-desktop@just-perfection/CHANGELOG.md
@@ -6,6 +6,24 @@ We go to the next version after each release on [GNOME Shell Extensions website]
 
 ## [Unreleased]
 
+## [19.0.0 Ancher] - 2022-03-02
+
+### Fixed
+
+- Blurry search entry on GNOME Shell themes with box-shadow.
+- Prefs file chooser recursion.
+- SecondaryMonitorDisplay error on GNOME Shell 42.
+- Shell theme override OSD for GNOME Shell 42.
+- Shell theme override workspace switcher for GNOME Shell 42.
+- Workspace popup visiblity in GNOME Shell 42.
+
+### Added
+
+- Libadwaita for GNOME Shell 42 prefs.
+- Panel icon size.
+- Panel world clock visiblity.
+- Weather visiblity.
+
 ## [18.0.0 Roslin] - 2022-02-12
 
 ### Fixed
@@ -332,4 +350,5 @@ We go to the next version after each release on [GNOME Shell Extensions website]
 - Hide top panel.
 - Hide workspace switcher.
 - Hot corner to toggle overview visibility.
+